requirements.txt
python-dotenv
requests
pandas
numpy

openai
tiktoken

chromadb
faiss-cpu

beautifulsoup4
lxml
feedparser

streamlit

plotly
altair

fastapi
uvicorn

spacy

python-dateutil
httpx

pydantic
